# fix: faster cache key factory for range