@H4ad 2023-04-06
@ljharb You have an interesting point, if the `parseOptions` was written just because of this function, so yes, maybe the necessity of parsing the options is just needed to check if is not undefined/null or bollean, In the other cases, we can just use the same object.

About the perf, I think it will not affect us since we already look for properties in the object that could doesn't exist, and if we just perform `Object.freeze` for an empty object and `loose`, we still save memory and avoid doing more comparisons and reduce the maintenance.