@wraithgar 2023-04-06
`rtl` is only used in `functions/coerce.js`, a standalone function not reused in any other part of semver.

This refactor is great because our coverage tests are exposing the fact that we are evaluating for a parameter that doesn't matter to this class.  The flag doesn't affect the data being memoized so it should not need to be accounted for in the memo key.