@wraithgar 2023-04-06
parseOptions is doing two things, it is allowing for the `options` parameter to overload as a boolean representing `loose`, and it is also ensuring that the object keys are in a deterministic order.

I searched through the code to see if `options` was doing any other heavy lifting and found:
 - `classes/comparator.js#intersects` recreating the logic instead of calling `parseOptions`
 - `functions/inc.js` making the positional parameter `options` optional by testing if it is a string,  this likely would only be affected when interpreting params from the cli, which already builds up an options object and only uses that.  I don't think we need to support passing `'true'` for the `loose` paramater.

Nothing I found would indicate that the ordering of parseOptions matters except in this one case where we are using it to memoize a cache key.  I think that memo logic should be isolated to where it's used, and made to only account for flags that affect the results.
 - `ranges/subset.js` looks at `options.includePrerelease` without having cleaned it via `parseOptions`