@wraithgar 2023-04-06
So far so good imho.  I am going to ping @kurtextrem and @jakebailey since they were pretty well engaged with the other performance PR.  If either of you don't want me to ping you on this and the next performance PR(s) please let me know.