@jakebailey 2023-04-06
Always happy to be pinged 😃 

I'll test these PRs out on my super stressful DefinitelyTyped monorepo case once I'm at my desk.