@H4ad 2023-04-06
> * `classes/comparator.js#intersects` recreating the logic instead of calling `parseOptions`

I will push a fix to this.

> * `functions/inc.js` making the positional parameter `options` optional by testing if it is a string,  this likely would only be affected when interpreting params from the cli, which already builds up an options object and only uses that.  I don't think we need to support passing `'true'` for the `loose` parameter.

I think the logic behind the testing is when we send the `identifier` on `options`, like:

```ts
inc(version: string, release: string, options: any, identifier: string);
inc(version: string, release: string, identifier: string);
```

> * `ranges/subset.js` looks at `options.includePrerelease` without having cleaned it via `parseOptions`

I already fix this case in this PR by calling `parseOptions` in the beginning.

> Nothing I found would indicate that the ordering of parseOptions matters except in this one case where we are using it to memoize a cache key. I think that memo logic should be isolated to where it's used, and made to only account for flags that affect the results.

In this case, should I remove all the object freeze from the other PR and change the tests to accept any object?

> If those options affect what classes/range.js returns there will need to be tests that cover those lines, which will fail if we are improperly memoizing the results.

So currently we don't need any additional test, right?