@wraithgar 2023-04-06
> this is much simpler and thus much less risky, well done

Agreed.  The past two weeks @H4ad has been extremely patient and diligent responding to all of the feedback in the PRs they've made.  The end results have been worth it and I am thankful for the effort being made here.

I've given this a 👍 and will let it sit now through the weekend to give others time to catch up.