@jakebailey 2023-04-06
Here's my DT stress test using pnpm. Before:

```
Done in 2m 14.5s
total time:  134.82s
user time:   162.81s
system time: 31.61s
CPU percent: 144%
max memory:  1752 MB
```

This PR (`"semver": "github:h4ad-forks/node-semver#fix/better-cache-key-factory"`):

```
Done in 2m 13s
total time:  133.25s
user time:   161.42s
system time: 31.23s
CPU percent: 144%
max memory:  1768 MB
```

So, seemingly no impact. This is probably not showing much improvement becuase I went and added a Range cache to pnpm (which yarn berry also has). I can try removing that cache and show before/after this PR, though, if people are interested.