@jakebailey 2023-04-06
Eh, I just ran it anyway.

Here's pnpm without the Range cache, instead using `semver.satisfies` like it used to:

```
Done in 2m 43s
total time:  163.27s
user time:   194.27s
system time: 31.76s
CPU percent: 138%
max memory:  1792 MB
```

Already a lot slower than my baseline. With this PR, it gets a little better:

```
Done in 2m 37.5s
total time:  157.75s
user time:   189.12s
system time: 31.90s
CPU percent: 140%
max memory:  1771 MB
```

Which, is still a light improvement, but not game changing like the options PR for pnpm anyway. Not to say this isn't a good change; absolutely take it 😄