@wraithgar 2023-04-06
Thanks @jakebailey it's good to have benchmarks that use a whole program.  Sometimes an optimization helps a use case that isn't very common.

The primary benefit of this PR is that it sets up the parse options refactor, since the memoization is now decoupled from that function.