@wraithgar 2023-04-10
@H4ad looks like https://github.com/npm/node-semver/pull/530 conflicted w/ this PR.  Once that is resolved I'll land this.